pyproject.toml
[project]
name = "self-aware-adaptive-retrieval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "datasets",
    "ratelimit",
    "openai",
    "python-dotenv",
]
